feat: add mapper for data models, finish FoxPhotoRepositoryImpl.kt
