+++ app/src/main/java/com/github/taasonei/randomfox/data/source/local/LastFoxPhotoDataSource.kt
@@ -5,7 +5,7 @@ import com.github.taasonei.randomfox.domain.model.FoxPhoto
 
 interface LastFoxPhotoDataSource {
 
-    suspend fun getLastFox(): FoxPhoto
+    suspend fun getLastFox(): FoxPhoto?
 
     suspend fun upsert(lastFox: LastFox)
